Remove ResizeObserver lifecycle DCHECK
LocalFrameView::NotifyResizeObservers had a DCHECK that the lifecycle
was clean, but this can be violated if an earlier resize observer (from
a different frame) dirtied the tree. We can safely remove this DCHECK
and continue for the same reasons that we continue when a modification
occurs in the presence of two same-frame resize observers.

Bug: 1368458
Change-Id: I5d486d9194e311f1d71ae94e4f688eccbc744fd2