Add notifyEvent API for fenced frames.
This API allows DOM events that occur within a fenced frame
to be observed by the frame's embedder in a privacy-respecting manner.
When an eligible event type (currently, just click) occurs within
the embedded document, window.fence.notifyEvent(e) can be called
with the event object, which will create a censored version of
the event to be handled by the embedder via addEventListener().

This API is being implmented as part of the "fenced frames with
local unpartitioned data access" proposal, see explainer here:
https://github.com/WICG/fenced-frame/blob/master/explainer/fenced_frames_with_local_unpartitioned_data_access.md

Design doc for this CL is here:
https://docs.google.com/document/d/1EiysMPnARSx2pzI1hSeFlI7k2ETxs5ycZoV2Cy08PEk/edit?usp=sharing

Bug: 1515579
Change-Id: I7f155e5640e7b798ca33a48c6afdda9c039d45c0