serial: Expose connection state for Bluetooth serial ports
Relevant spec change: https://github.com/WICG/serial/pull/197

This CL adds a connected attribute to the SerialPort interface
that exposes the logical connection state of the serial port.
The behavior for Bluetooth serial ports is changed so that the
connect and disconnect events are dispatched when the connection
state of the Bluetooth device changes. The behavior is unchanged
for wired serial ports.

Intent to Prototype: TODO

Bug: 1488031
Change-Id: I34b88246e62202c24eee08ad4a997aff7eadeebb