Fenced frames: Make same-origin fenced frame redaction check consistent
Fix at least one potential cause of a BadMessage where renderer's view
of "same-origin to mapped url" and browser's view of "same-origin to
mapped url" differ.

The origin to be committed is not necessarily the origin of the url to
be committed, e.g. for sandboxed iframes. All our browser-side checks
are with respect to the last committed origin, so we should use the
origin to be committed when redacting fenced frame properties for the
renderer.

Bug: 1515202, 1508030
Change-Id: I4e1bfde4c0bab7c88a93ae0f01d8568ead3fdc18