Do not intercept for prefetch in subframes.
Other parts of the prefetch code assume this is for top-level
navigation. If we do want to support this, we'd have to do
adjust it to properly handle cookie and cache partitioning
in a more sophisticated way.

The corresponding test that uses iframes is removed; there is
already parallel coverage using window.open. We could in
theory disable it instead, but a test that runs nowhere seems
less useful and we can always find it in git history if we
do start supporting this in some way.

Bug: 1511780
Change-Id: I249432e6a5de4b6ffbf1d7c6c260a78850151d3b