Fix Highlight Pseudos using container-relative units
Add support for creating a new style when the highlight uses container-relative units and has a different container to that of its parent.

Add testing of highlights using container queries.

Fixes a typo too.

Units depending on block/inline direction still do
not work for vertical writing modes. That's for another CL.

Bug: 1468306
Change-Id: I781a35810133f3591c7a3587ff5d13df0dfac9ee